move swap helpers back to useSwapQuote.ts
